requirements.txt
streamlit
google-generativeai
st-click-detector
watchdog
